refactor(domain): BaseEntity audit, long ids, and feature entity folders
Restructure Domain into Common/Entities base plus Todos/Entities and Users/Entities.
Apply Created/Updated stamps in SaveChanges with IAuditUserProvider. Migrate
identifiers to bigint; align contracts, Jwt, CurrentUserService, and rules.

+++ src/ToDoList.Api/Security/CurrentUserService.cs
@@ -5,8 +5,8 @@ namespace ToDoList.Api.Security;
 
 public sealed class CurrentUserService(IHttpContextAccessor accessor) : ICurrentUserService
 {
-    public int UserId =>
-        int.TryParse(
+    public long UserId =>
+        long.TryParse(
             accessor.HttpContext?.User.FindFirstValue(ClaimTypes.NameIdentifier),
             out var id)
             ? id
