+++ src/ToDoList.Application/Abstractions/ICurrentUserService.cs
@@ -2,5 +2,5 @@ namespace ToDoList.Application.Abstractions;
 
 public interface ICurrentUserService
 {
-    int UserId { get; }
+    long UserId { get; }
 }
