+++ src/ToDoList.Application/Todos/Commands/Create/CreateTodoCommandHandler.cs
@@ -1,14 +1,14 @@
 using MediatR;
 using ToDoList.Application.Abstractions;
 using ToDoList.Contracts.Todos.Commands.Create;
-using ToDoList.Domain.Entities;
+using ToDoList.Domain.Todos.Entities;
 
 namespace ToDoList.Application.Todos.Commands.Create;
 
 public sealed class CreateTodoCommandHandler(IAppDbContext db, ICurrentUserService currentUser)
-    : IRequestHandler<CreateTodoCommand, int>
+    : IRequestHandler<CreateTodoCommand, long>
 {
-    public async Task<int> Handle(CreateTodoCommand request, CancellationToken cancellationToken)
+    public async Task<long> Handle(CreateTodoCommand request, CancellationToken cancellationToken)
     {
         var userId = currentUser.UserId;
         var todo = new Todo(userId, request.Title.Trim());
