+++ src/ToDoList.Contracts/Auth/Commands/Register/RegisterCommand.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace ToDoList.Contracts.Auth.Commands.Register;
 
-public sealed record RegisterCommand(string Username, string Password) : IRequest<int>;
+public sealed record RegisterCommand(string Username, string Password) : IRequest<long>;
