+++ src/ToDoList.Contracts/Todos/Commands/Create/CreateTodoCommand.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace ToDoList.Contracts.Todos.Commands.Create;
 
-public sealed record CreateTodoCommand(string Title) : IRequest<int>;
+public sealed record CreateTodoCommand(string Title) : IRequest<long>;
