+++ src/ToDoList.Contracts/Todos/Commands/Delete/DeleteTodoCommand.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace ToDoList.Contracts.Todos.Commands.Delete;
 
-public sealed record DeleteTodoCommand(int TodoId) : IRequest;
+public sealed record DeleteTodoCommand(long TodoId) : IRequest;
